requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-environ==0.9.0
django-filter==21.1
django-autoslug==1.9.8
Pillow==8.3.2
django-resized==1.0.2
psycopg2-binary==2.9.1
flake8==3.9.2
black==21.6b0
isort==5.9.3
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
appdirs==1.4.4
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
django-countries==7.2.1
django-rest-auth==0.9.2
django-templated-mail==1.1.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
pathspec==0.11.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
python3-openid==3.2.0
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.14
